+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### [0.3.0-b11] Preparing the house for a new core.
-. `NEW` Added a minimalistic logger.
+- `FIX` Improved the logger with a timestamp, caller name, and filters.
+- `NEW` Added a minimalistic logger.
 - `NEW` Added unresolved test to make sure enumerable properties are really enumerable.
 - `FIX` Converted constants on main instantiator to upper case.
 - `NEW` Added optional parameters on main instance's default methods.